mcp-server-sample/pyproject.toml
[project]
name = "mcp-server-sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.1.0",
]
